gerenciamento de caminhoes / motoristas

+++ src/main/java/com/ja/chegou/ja_chegou/service/DistributionCenterServiceImpl.java
@@ -2,6 +2,8 @@ package com.ja.chegou.ja_chegou.service;
 
 import com.ja.chegou.ja_chegou.entity.DistributionCenter;
 import com.ja.chegou.ja_chegou.repository.DistributionCenterRepository;
+import com.ja.chegou.ja_chegou.repository.RouteRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,6 +11,8 @@ import java.util.List;
 @Service
 public class DistributionCenterServiceImpl implements DistributionCenterService{
 
+     @Autowired
+    RouteRepository routeRepository;
 
     private final DistributionCenterRepository distributionCenterRepository;
 
@@ -26,8 +30,11 @@ public class DistributionCenterServiceImpl implements DistributionCenterService{
         return distributionCenterRepository.save(center);
     }
 
-    @Override
     public void delete(Long id) {
+        if (!routeRepository.findByOriginId(id).isEmpty()) {
+            throw new IllegalStateException("Não é possível excluir: existem rotas vinculadas a esta central.");
+        }
         distributionCenterRepository.deleteById(id);
     }
+
 }
